update2

+++ Admin/app/src/main/java/com/example/admin/pojo/Order.kt
@@ -4,7 +4,7 @@ import com.example.admin.utils.OrderState
 import kotlin.properties.Delegates
 
 /** only one constructor because this is admin app, it receives orders only*/
-open class Order {
+open class Order{
     var id by Delegates.notNull<String>()     // identifier of order obj
         private set
     var client by Delegates.notNull<Client>()
@@ -15,7 +15,7 @@ open class Order {
         private set
     var items by Delegates.notNull<MutableList<OrderItem>>()
         private set
-    var state: Byte = OrderState.ACCEPTED
+    var state: Int = OrderState.ACCEPTED
         protected set
 
     /**
