+++ Admin/app/src/main/java/com/example/admin/utils/OrderState.kt
@@ -2,7 +2,7 @@ package com.example.admin.utils
 
 class OrderState {
     companion object {
-        const val ACCEPTED: Byte = 0b0
-        const val CANCELED: Byte = 0b1
+        const val ACCEPTED: Int = 0b0
+        const val CANCELED: Int = 0b1
     }
 }
